adding comments, changing mark as favorite button, adding remember me functionality
