+++ models/index.js
@@ -1,7 +1,6 @@
 const User = require('./User');
 const Recipe = require('./Recipe');
 
-
 // this file determines the relationship between each model
 
 // demonstrating the relationship between users and recipes
@@ -11,9 +10,6 @@ User.hasMany(Recipe, {
 Recipe.belongsTo(User, {
     foreignKey: 'user_id',
 });
-
-
-
 
 
 module.exports = { User, Recipe };
