+++ public/js/login.js
@@ -2,23 +2,45 @@ const submitSignupBtn = $('#submit-sign-up-btn');
 const submitLoginBtn = $('#submit-login-btn');
 const logOutBtn = $('#logout-btn');
 
+// this checks if the user has their login information in local storage
+function init() {
+
+    // getting information from local storage
+    const email = JSON.parse(localStorage.getItem("email"))
+    const password = JSON.parse(localStorage.getItem("password"))
+    const checked = JSON.parse(localStorage.getItem("checked"))
+
+    // if the user has chosen to be remembered then their infromation in autofilled. If not then local storage is cleared
+    if (checked) {
+        $('#login-email').val(email)
+        $('#login-password').val(password)
+        $('#remember-me').prop("checked", true);
+    } else {
+        localStorage.clear();
+    }
+}
+
 // this function allows a new user to sign up
 const signupNewUser = async (e) => {
     e.preventDefault();
 
+    // getting input from the different text areas
     const rawUserName = $('#new-account-name');
     const rawEmail = $('#new-account-email');
     const rawPassword = $('#new-account-password');
 
+    // turning that input into usable data
     const users_name = rawUserName.val().trim();
     const email = rawEmail.val().trim();
     const password = rawPassword.val().trim();
 
+    // this checks if the password the user inputed is at least 8 characters long. If it is not the user is alerted
     if (password.split("").length < 8) {
         alert("password must be at least 8 characters long. Please try again");
         return;
     }
 
+    // if all of the input fields have information in them then the information is sent to the server to be stored in a database
     if (users_name && email && password) {
         const response = await fetch('/api/users', {
             method: 'POST',
@@ -26,6 +48,7 @@ const signupNewUser = async (e) => {
             headers: { 'Content-Type': 'application/json' },
         });
 
+        // if the information is saved then the user is taken back to the homepage
         if (response.ok) {
             document.location.replace('/');
         } else {
@@ -41,9 +64,25 @@ const signupNewUser = async (e) => {
 const logUserIn = async (e) => {
     e.preventDefault();
 
-    const email = $('#login-email').text.trim();
-    const password = $('#login-password').text.trim();
+    // getting input from the different text areas
+    const rawEmail = $('#login-email');
+    const rawPassword = $('#login-password');
 
+    // turning that input into usable data
+    const email = rawEmail.val().trim();
+    const password = rawPassword.val().trim();
+
+    // this checks to see if the remember me checkbox is checked. If it is then the users information is saved into local storage
+    if ($('#remember-me').is(':checked')) {
+        localStorage.setItem("email", JSON.stringify(email));
+        localStorage.setItem("password", JSON.stringify(password));
+        localStorage.setItem("checked", JSON.stringify(true));
+
+    } else {
+        localStorage.setItem("checked", JSON.stringify(false));
+    }
+
+    //   if there is both an email and password they are sent to the server to check the database
     if (email && password) {
         const response = await fetch('/api/users/login', {
             method: 'POST',
@@ -51,6 +90,7 @@ const logUserIn = async (e) => {
             headers: { 'Content-Type': 'application/json' },
         });
 
+        // if the credintials are corret the user is then logged in and they are sent to the recipe page
         if (response.ok) {
             document.location.replace('/recipe');
         } else {
@@ -59,19 +99,19 @@ const logUserIn = async (e) => {
     }
 };
 
-
+// this logs a user out
 const logoutUser = async (e) => {
     e.preventDefault();
-
     const response = await fetch('/api/users/logout', {
         method: 'POST',
         body: JSON.stringify(),
         headers: { 'Content-Type': 'application/json' },
     });
     document.location.replace('/');
-
 }
 
+// starts on page load
+init()
 
 // event listeners to trigger each function
 submitSignupBtn.on('click', signupNewUser);
