+++ seeds/userSeed.json
@@ -1,7 +1,7 @@
 [
     {
         "users_name": "user1",
-        "email": "example@email.com",
+        "email": "example@emai2l.com",
         "intolerances": "",
         "password": "123456789"
     },
