+++ views/layouts/main.handlebars
@@ -81,9 +81,9 @@
     <script src="https://unpkg.com/bulma-modal-fx/dist/js/modal-fx.min.js"></script>
     <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"></script>
     <script src='https://cdnjs.cloudflare.com/ajax/libs/overlayscrollbars/1.9.1/js/OverlayScrollbars.min.js'></script>
-    <script src="/js/recipeHTML.js"></script>
+    <script src="/js/recipe.js"></script>
     <script src="/js/login.js"></script>
-    <script src="/js/landingHTML.js"></script>
+    <script src="/js/homepage.js"></script>
 
 
     </body>
