+++ views/login.handlebars
@@ -51,6 +51,10 @@
                     <div class="control">
                         <button class="button is-link" id="submit-login-btn"=>Welcome Back</button>
                     </div>
+                    <div class="control">
+                        <input type="checkbox" id="remember-me">
+                        <label for="remember-me">Remember me</label>
+                    </div>
                 </div>
             </div>
