+++ views/partials/new-column.handlebars
@@ -16,9 +16,11 @@
                 <div class="btn-container"><button class="button save-recipe"><i>Save</i></button><button
                         class="button favorite-btn"><i>
                             {{#if favorite}}
-                            ⭐
+                            ⭐ Remove Favorite
+                            {{else}}
+                            Mark as Favorite
                             {{/if}}
-                            Mark as Favorite</i></button><button class="button remove-recipe-btn"><i>Remove
+                        </i></button><button class="button remove-recipe-btn"><i>Remove
                             recipe</i></button>
                 </div>
             </div>
