--- views/signup.handlebars
@@ -1,30 +1,0 @@
-<main class="custom-content-area pt-4">
-
-
-    <div class="custom-card mx-5">
-
-        <div class="custom-card-header d-block p-2">
-            <h4 class="card-title ">Sign Up</h4>
-            <div style="clear: both;"></div>
-        </div>
-
-        <div class="custom-card-body d-block">
-            <div class="form-container ">
-                <form>
-                    <div class="form-group d-block my-1 mx-2">
-                        <label for="username">Username</label>
-                        <textarea class="form-control" id="username" rows="1"></textarea>
-                    </div>
-                    <div class="form-group my-1 mx-2">
-                        <label for="password">Password</label>
-                        <input type="password" class="form-control" id="password" placeholder="Password">
-                    </div>
-                    <button class="btn my-1 mx-2" id="submit-sign-up-btn">Submit</button>
-                </form>
-                <a class="login-redirect-link mx-2 mb-3 d-block" href="/login">Sign in instead</a>
-            </div>
-        </div>
-
-    </div>
-
-</main>
